% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_BroadBand.R
\name{Get_BroadBand}
\alias{Get_BroadBand}
\title{Download the data regarding the broad band connection activation in Italian schools}
\source{
\href{https://bandaultralarga.italia.it/scuole-voucher/dashboard-scuole/}{Broadband dashboard}
}
\usage{
Get_BroadBand(
  Date = Sys.Date(),
  verbose = TRUE,
  show_col_types = FALSE,
  autoAbort = FALSE
)
}
\arguments{
\item{Date}{Object of class \code{Date}. The date at which it is required to determine if the broad band connection has been activated or not. By default it is the current date.}

\item{verbose}{Logical. If \code{TRUE}, the user keeps track of the main underlying operations. \code{TRUE} by default.}

\item{show_col_types}{Logical. If \code{TRUE}, if the \code{verbose} argument is also \code{TRUE}, the columns of the raw dataset are shown during the download. \code{FALSE} by default.}

\item{autoAbort}{Logical. Whether to automatically abort the operation and return NULL in case of missing internet connection or server response errors. \code{FALSE} by default.}
}
\value{
An object of class \code{tbl_df}, \code{tbl} and \code{data.frame}. The variables \code{BB_Activation_date} and \code{BB_Activation_staus} indicate the activation date and activation status of the broadband connection at the selected date.
}
\description{
Retrieves the data regarding the activation date of the broad band connection in schools. It also indicates whether the connection was activated or not at a certain date.
}
\details{
Ultra - Broadband is defined as everlasting internet connection with a
maximum speed of 1 gigabit per second, with a minimum guaranteed speed of
100 megabits/second both on the uploading and downloading operations, until
the peering point is reached, as declared on the data provider's \href{https://bandaultralarga.italia.it/scuole-voucher/progetto-scuole/}{website}.
In the example the broadband availability at the beginning of school  year 2022/23 (1st september 2022) is shown.
}
\examples{

\donttest{
Broadband_220901 <- Get_BroadBand(Date = as.Date("2022-09-01"), autoAbort = TRUE)

Broadband_220901

Broadband_220901[, c(9,6,13,14)]
}




}
