% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchoolDataIT-package.R
\docType{package}
\name{SchoolDataIT-package}
\alias{SchoolDataIT}
\alias{SchoolDataIT-package}
\title{SchoolDataIT: Retrieve, Harmonise and Map Open Data Regarding the Italian School System}
\description{
Compiles and displays the available data sets regarding the Italian school system, with a focus on the infrastructural aspects. Input datasets are downloaded from the web, with the aim of updating everything to real time. The functions are divided in four main modules, namely 'Get', to scrape raw data from the web 'Util', various utilities needed to process raw data 'Group', to aggregate data at the municipality or province level 'Map', to visualize the output datasets.
}
\author{
\strong{Maintainer}: Leonardo Cefalo \email{leonardo.cefalo@uniba.it} (\href{https://orcid.org/0009-0001-7776-723X}{ORCID})

Other contributors:
\itemize{
  \item Alessio Pollice \email{alessio.pollice@uniba.it} (\href{https://orcid.org/0000-0002-2818-9373}{ORCID}) [contributor, thesis advisor]
  \item Paolo Maranzano \email{pmaranzano.ricercastatistica@gmail.com} (\href{https://orcid.org/0000-0002-9228-2759}{ORCID}) [contributor]
}

}
\keyword{internal}
